mcp-memex/pyproject.toml
[project]
name = "mcp-memex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "firecrawl-py>=1.6.3",
    "ipykernel>=6.29.5",
    "mcp>=1.1.0",
    "numpy>=2.1.3",
    "pymilvus[model]>=2.5.0",
    "python-dotenv>=1.0.1",
    "voyageai>=0.3.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
mcp-memex = "mcp_memex:main"
